pyproject.toml
[tool.poetry]
name = "django-graphql-subscription"
version = "0.1.0"
description = ""
authors = ["Patrick <patrickckabwe@gmail.com>"]
readme = "README.md"
packages = [{include = "django_graphql_subscription"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
strawberry-graphql-django = "^0.8.2"
strawberry-graphql = {extras = ["channels"], version = "^0.151.0"}
uvicorn = {extras = ["standard"], version = "^0.20.0"}
psycopg = {extras = ["binary"], version = "^3.1.5"}
psycopg2-binary = "^2.9.5"
django-environ = "^0.9.0"
flake8 = "^6.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn config.asgi:application --reload --host 0.0.0.0"
